backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
groq==0.4.0
pinecone-client==2.2.4
numpy==1.21.0
pandas==1.3.0
requests==2.26.0
python-multipart==0.0.5
pydantic==1.8.2
typing-extensions==3.10.0.0 